requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
fonttools==4.32.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
protobuf==4.21.1
PyMySQL==1.0.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.0.3
